refactor: character List component
